+++ src/components/LessonEditor.tsx
@@ -3,13 +3,18 @@ import { ModalContent } from "./Modals";
 import Button from "./basic/button";
 import Input from "./basic/input";
 import TextArea from "./basic/textarea";
-import { MdInfo } from "react-icons/md";
+import { MdExpandLess, MdExpandMore, MdInfo } from "react-icons/md";
 import { getUserId } from "@auth";
 import { upsertLesson, useLessons } from "@db/lesson";
+import Checkbox from "./basic/checkbox";
 
 export default function LessonEditorModal(): JSX.Element | null {
   const lessonEditor = useLessonEditorStore();
   const lesson = lessonEditor.lesson;
+  const delimiter =
+    (lesson?.settings?.delimiter?.enabled &&
+      lesson?.settings?.delimiter?.value) ||
+    " ";
 
   if (!lessonEditor.isEditorOpen) {
     return null;
@@ -80,15 +85,79 @@ export default function LessonEditorModal(): JSX.Element | null {
       <TextArea
         className="mt-5 grow"
         label="Words, n-grams to practice. Separated by spaces by default"
-        value={lesson.words.join(" ")}
+        value={lesson.words.join(delimiter)}
         placeholder="Type words here..."
         disabled={!lesson.custom}
         onChange={(e) => {
-          lesson.words = e.currentTarget.value.split(" ");
+          lesson.words = e.currentTarget.value.split(delimiter);
           lessonEditor.setLesson(lesson);
         }}
       />
+      <AdvancedSettings />
     </ModalContent>
+  );
+}
+
+function AdvancedSettings(): JSX.Element {
+  const {
+    isAdvancedOpen,
+    setAdvancedOpen,
+    lesson,
+    setDelimiter,
+    setKeepSpace,
+  } = useLessonEditorStore();
+
+  if (!isAdvancedOpen) {
+    return (
+      <div
+        className="mt-5 flex items-center cursor-pointer"
+        onClick={() => setAdvancedOpen(true)}
+      >
+        <MdExpandMore className="w-8 h-8" />
+        Advanced Settings
+      </div>
+    );
+  }
+
+  return (
+    <div className="mt-5">
+      <div
+        className="flex items-center cursor-pointer"
+        onClick={() => setAdvancedOpen(false)}
+      >
+        <MdExpandLess className="w-8 h-8" />
+        <span>Advanced Settings</span>
+      </div>
+      <div className="mt-5">
+        <Checkbox
+          label="Delimiter"
+          disabled={!lesson.custom}
+          helpText="Use another delimiter than space, can be multiple characters. Useful when you want to use words or symbols that contains spaces."
+          checked={!!lesson?.settings?.delimiter?.enabled}
+          onChecked={(checked) =>
+            setDelimiter(checked, lesson?.settings?.delimiter?.value)
+          }
+        >
+          Delimiter:
+          <input
+            className="ml-2 w-10 text-black"
+            name="delimiter"
+            disabled={!lesson.custom || !lesson?.settings?.delimiter?.enabled}
+            value={lesson?.settings?.delimiter?.value}
+            onChange={(e) => setDelimiter(true, e.currentTarget.value)}
+          />
+        </Checkbox>
+      </div>
+      <div className="mt-5">
+        <Checkbox
+          label="Keep spaces"
+          disabled={!lesson.custom}
+          helpText="Don't remove leading and trailing spaces from words. For example, combining ' == ' with ' != ' will keep two spaces between the words."
+          checked={!!lesson?.settings?.keepSpaces}
+          onChecked={(checked) => setKeepSpace(checked)}
+        />
+      </div>
+    </div>
   );
 }