+++ src/components/basic/checkbox.tsx
@@ -2,6 +2,7 @@ export interface CheckboxProps {
   label: string;
   checked: boolean;
   disabled?: boolean;
+  helpText?: string;
   onChecked: (checked: boolean) => void;
   children?: React.ReactNode;
 }
@@ -11,6 +12,7 @@ export default function Checkbox({
   checked,
   disabled,
   children,
+  helpText,
   onChecked,
 }: CheckboxProps): JSX.Element {
   return (
@@ -26,6 +28,12 @@ export default function Checkbox({
         onChange={(checked) => onChecked(checked.currentTarget.checked)}
       />
       {children ? children : label}
+      {helpText && (
+        <>
+          <br />
+          <span className="text-sm text-gray-400">{helpText}</span>
+        </>
+      )}
     </div>
   );
 }