+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 0.2.0 - 2024-07-05
 
 ### ➕ Added
 
 - Top menu with profile and login/logout menu
 - Logout functionality
+- Advanced settings for lessons
+  - Add custom delimiter for words. Allows you to add words with spaces.
+  - Keep spaces will not trim spaces at the end of the word (useful for practicing code)
 
 ### 🔄 Changed
 
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🐛 Fixed
 
 - Pressing space at the start does no longer clear the current WPM
+- Couldn't create a new lesson as it was treated as a built-in lesson
 
 ## 0.1.0 - 2024-07-03